chore: added UUID verification before manipulate db

+++ src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactsRepository = require("../repositories/ContactsRepository");
+const isValidUUID = require("../utils/isValidUUID");
 
 class ContactController {
   async index(request, response) {
@@ -11,6 +12,10 @@ class ContactController {
   async show(request, response) {
     const { id } = request.params;
 
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: "Invalid user ID" });
+    }
+
     const contact = await ContactsRepository.findById(id);
 
     if (!contact) {
@@ -21,25 +26,30 @@ class ContactController {
   }
 
   async store(request, response) {
-    const {
-      name, email, phone, category_id
-    } = request.body;
+    const { name, email, phone, category_id } = request.body;
 
     if (!name) {
       return response.status(400).json({ error: "Name is required" });
     }
 
-    const contactExists = await ContactsRepository.findByEmail(email);
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: "Invalid category" });
+    }
 
-    if (contactExists) {
-      return response.status(400).json({ error: "This Email already in use" });
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+      if (contactExists) {
+        return response
+          .status(400)
+          .json({ error: "This Email already in use" });
+      }
     }
 
     const contact = await ContactsRepository.create({
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
     response.status(201).json(contact);
@@ -47,29 +57,39 @@ class ContactController {
 
   async update(request, response) {
     const { id } = request.params;
-    const {
-      name, email, phone, category_id
-    } = request.body;
+    const { name, email, phone, category_id } = request.body;
 
-    const contactExists = await ContactsRepository.findById(id);
-    if (!contactExists) {
-      return response.status(404).json({ error: "Contact doesn't exists" });
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: "Invalid user ID" });
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: "Invalid category" });
     }
 
     if (!name) {
       return response.status(400).json({ error: "Name is required" });
     }
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id !== id) {
-      return response.status(400).json({ error: "This Email already in use" });
+    const contactExists = await ContactsRepository.findById(id);
+    if (!contactExists) {
+      return response.status(404).json({ error: "Contact doesn't exists" });
+    }
+
+    if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response
+          .status(400)
+          .json({ error: "This Email already in use" });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
     response.json(contact);
@@ -78,6 +98,9 @@ class ContactController {
   async delete(request, response) {
     const { id } = request.params;
 
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: "Invalid user ID" });
+    }
     await ContactsRepository.delete(id);
     response.sendStatus(204);
   }
